Move applyReplacements to :utils
