+++ gradle-palantir-java-format/src/main/java/com/palantir/javaformat/gradle/PalantirJavaFormatPlugin.java
@@ -16,12 +16,10 @@
 
 package com.palantir.javaformat.gradle;
 
-import com.palantir.javaformat.java.FormatDiff;
 import java.io.IOException;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.tasks.TaskAction;
 
 public final class PalantirJavaFormatPlugin implements Plugin<Project> {
